+++ src/examples/types/User.ts
@@ -1,0 +1,4 @@
+export type User = {
+	userId: number;
+	name: string;
+};